Use size_t for qname value length and unsigned int for name lengths.
raptor_qname_s changes value_length to size_t

(raptor_new_qname, raptor_new_qname_from_namespace_local_name,
raptor_qname_string_to_uri): unsigned ints for prefix and name
lengths.

+++ src/raptor_internal.h
@@ -787,7 +787,7 @@ struct raptor_qname_s {
   raptor_uri *uri;
   /* optional value - used when name is an attribute */
   const unsigned char *value;
-  unsigned int value_length;
+  size_t value_length;
 };
